Revert removal of the patch versions (#44)
This readds the patch versions in the cargo version specifiers to ensure
that we include a tested working version. In theory there could have been
breakage with one of the included patch versions of crates we depend on
so it is best practice to supply a full version specifier. Cargo will
automatically pull in a higher version when running `cargo update` or
building this library with another crate that requires a higher version.

See https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html#specifying-dependencies-from-cratesio

This reverts commit 313eac446c27b30dc12013692d770d37cd38522f.