Use chaining for `AnsiGenericString::hyperlink` (#48)
The API for modifying a string to contain a hyperlink used in #47
behaves different than most of our `Style`-driven APIs that basically
all work through method chaining.

As an alternative I propose we change the signature to take the
`AnsiGenericString` by value and modify it in place. This feels more
expressive for quickly building out a styled string with a link.
One downside compared to a modifying setter API is that it takes a
more tokens to do conditional modification with an `if` expression.